Exclusively parse with offsets not straight thru
Read the end record which has the offset of the central directory, then
use the local header offset in each central directory to read local
headers. This is the canonical way to parse zip contents. For fun, I can
allow reading local file headers straight through, which might reveal
"lost" files not in the central directory.